+++ BuiltInChipsSource/Add16.java
@@ -27,6 +27,6 @@ public class Add16 extends BuiltInGate {
     protected void reCompute() {
         int a = inputPins[0].get();
         int b = inputPins[1].get();
-        outputPins[0].set((int)(a + b));
+        outputPins[0].set(a + b);
     }
 }
