+++ BuiltInChipsSource/DMux.java
@@ -28,7 +28,7 @@ public class DMux extends BuiltInGate {
     protected void reCompute() {
         int in = inputPins[0].get();
         int sel = inputPins[1].get();
-        outputPins[0].set((int)(sel == 0 ? in : 0));
-        outputPins[1].set((int)(sel == 0 ? 0  : in));
+        outputPins[0].set(sel == 0 ? in : 0);
+        outputPins[1].set(sel == 0 ? 0  : in);
     }
 }
