+++ BuiltInChipsSource/DMux4Way.java
@@ -29,9 +29,9 @@ public class DMux4Way extends BuiltInGate {
     protected void reCompute() {
         int in = inputPins[0].get();
         int sel = inputPins[1].get();
-        outputPins[0].set((int)(sel == 0 ? in : 0));
-        outputPins[1].set((int)(sel == 1 ? in : 0));
-        outputPins[2].set((int)(sel == 2 ? in : 0));
-        outputPins[3].set((int)(sel == 3 ? in : 0));
+        outputPins[0].set(sel == 0 ? in : 0);
+        outputPins[1].set(sel == 1 ? in : 0);
+        outputPins[2].set(sel == 2 ? in : 0);
+        outputPins[3].set(sel == 3 ? in : 0);
     }
 }
