+++ BuiltInChipsSource/HalfAdder.java
@@ -29,7 +29,7 @@ public class HalfAdder extends BuiltInGate {
     protected void reCompute() {
         int a = inputPins[0].get();
         int b = inputPins[1].get();
-        outputPins[0].set((int)(a ^ b));
-        outputPins[1].set((int)(a & b));
+        outputPins[0].set(a ^ b);
+        outputPins[1].set(a & b);
     }
 }
